+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Changes
 * Improve timesheet calculation for shooting grounds in LS
+* Improve safety margins for shooting grounds in LS
 
 ## 2.1.1